Refactor non-alphanumeric into its own module
